Added error handling to connection pools for PostgreSQL
I was having issues where the application (sails.js in my case) would stop/crash due to uncaught error events being emitted by pg's pool system.

This was happening when the database is shutdown while a query is being actively run on it. I found the solution was to add these listeners for error events to the pools to catch and log them.

I was only getting an issue with the exec pool, but I added the same to the other pools for completeness. Feel free to alter the implementation or error messaging, I only need the error event to not be 'Unhandled' so the server doesn't stop. 

During testing I was still getting the normal errors coming through the promise rejections from the database so I hope this doesn't affect normal error generation.